+++ packages/yourloops/test/integration/pages/home-page/hcp-home-page.spec.tsx
@@ -29,19 +29,7 @@ import { act, fireEvent, screen, waitFor, within } from '@testing-library/react'
 import { loggedInUserId, mockAuth0Hook } from '../../mock/auth0.hook.mock'
 import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
-import {
-  flaggedPatient,
-  flaggedPatientId,
-  hypoglycemiaPatient,
-  monitoredPatient,
-  monitoredPatientTwo,
-  monitoredPatientWithMmol,
-  noDataTransferredPatient,
-  pendingPatient,
-  timeSpentOutOfTargetRangePatient,
-  unmonitoredPatient,
-  unreadMessagesPatient
-} from '../../data/patient.api.data'
+import { flaggedPatientId, monitoredPatient, pendingPatient, unmonitoredPatient } from '../../data/patient.api.data'
 import {
   buildAvailableTeams,
   buildFiltersTeam,
@@ -91,12 +79,9 @@ describe('HCP home page', () => {
 
   it('should not display the Care team tab and not allow to add patients if the private practice is selected', async () => {
     localStorage.setItem('selectedTeamId', 'private')
-    jest.spyOn(PatientApi, 'getPatientsForHcp').mockResolvedValueOnce([{
+    jest.spyOn(PatientApi, 'getScopedPatientsForHcp').mockResolvedValue([{
       ...monitoredPatient,
-      teams: [{
-        teamId: 'private',
-        status: UserInvitationStatus.accepted
-      }]
+      invitationStatus: UserInvitationStatus.accepted
     }])
 
     const router = renderPage('/')
@@ -181,13 +166,10 @@ describe('HCP home page', () => {
     expect(cancelButton).toBeEnabled()
     const confirmRemoveButton = within(removeDialog).getByRole('button', { name: 'Remove patient' })
 
-    jest.spyOn(PatientApi, 'getPatientsForHcp').mockResolvedValueOnce([monitoredPatient, monitoredPatientTwo, monitoredPatientWithMmol, pendingPatient])
-    const teamId = unmonitoredPatient.teams[0].teamId
     await act(async () => {
       await userEvent.click(confirmRemoveButton)
     })
-    expect(removePatientMock).toHaveBeenCalledWith(teamId, unmonitoredPatient.userid)
-    expect(within(dataGridCurrentRows).getAllByRole('row')).toHaveLength(4)
+    expect(removePatientMock).toHaveBeenCalledWith(myThirdTeamId, unmonitoredPatient.userid)
     expect(screen.queryByTestId('remove-hcp-patient-dialog')).toBeFalsy()
     expect(screen.getByTestId('alert-snackbar')).toHaveTextContent(`${unmonitoredPatient.profile.firstName} ${unmonitoredPatient.profile.lastName} is no longer a member of ${myThirdTeamName}`)
   })
@@ -196,7 +178,6 @@ describe('HCP home page', () => {
     localStorage.setItem('selectedTeamId', filtersTeamId)
     const teams = [buildFiltersTeam(), buildTeamThree(), buildPrivateTeam()]
     jest.spyOn(TeamAPI, 'getTeams').mockResolvedValue(teams)
-    jest.spyOn(PatientApi, 'getPatientsForHcp').mockResolvedValue([monitoredPatient, unreadMessagesPatient, timeSpentOutOfTargetRangePatient, hypoglycemiaPatient, noDataTransferredPatient, flaggedPatient, pendingPatient, monitoredPatientTwo])
     const router = renderPage('/')
     await waitFor(() => {
       expect(router.state.location.pathname).toEqual('/home')
@@ -205,7 +186,8 @@ describe('HCP home page', () => {
     await checkHCPLayout(`${firstName} ${lastName}`, { teamName: filtersTeamName }, teams)
     checkPatientListHeader()
 
-    expect(screen.queryByTestId('filters-label')).not.toBeInTheDocument()
+    expect(screen.getByTestId('filters-label')).toHaveTextContent('Filters deactivated: 6 patient(s) out of 6')
+    expect(screen.queryByTestId('reset-filters-link')).not.toBeInTheDocument()
     const dataGridRow = screen.getByTestId('patient-list-grid')
     expect(within(dataGridRow).getAllByRole('row')).toHaveLength(7)
     expect(dataGridRow).toHaveTextContent('PatientSystemTime spent out of the target rangeSevere hypoglycemiaData not transferredLast data updateActionsUnflag patient flagged@patient.frFlagged PatientDBLG110%20%30%N/ANo new messagesFlag patient hypoglycemia@patient.frHypoglycemia PatientDBLG110%20%30%N/ANo new messagesFlag patient monitored-patient@diabeloop.frMonitored PatientDBLG110%20%30%N/ANo new messagesFlag patient no-data@patient.frNo Data PatientDBLG110%20%30%N/ANo new messagesFlag patient time-out-of-range@patient.frTime Out of Range PatientDBLG110%20%30%N/ANo new messagesFlag patient unread-messages@patient.frUnread Messages PatientDBLG110%20%30%N/AThe patient has sent you new messagesData calculated on the last 7 daysRows per page:101–6 of 6')
@@ -221,6 +203,7 @@ describe('HCP home page', () => {
 
     // check the manual flag toggle
     await updatePatientsFilters({ ...defaultToggles, manualFlagFilterToggle: true })
+    expect(screen.getByTestId('filters-label')).toHaveTextContent('Filters activated: 1 patient(s) out of 6')
     checkDataGridAfterSinglePatientFilter(dataGridRow, 'Unflag patient flagged@patient.frFlagged PatientDBLG110%20%30%N/ANo new messages')
     await userEvent.click(filtersButton)
     checkPatientsFilters({ ...defaultToggles, manualFlagFilterToggle: true })
@@ -283,7 +266,7 @@ describe('HCP home page', () => {
 
     const filtersButtonDisabled = screen.getByRole('button', { name: 'Filters' })
     expect(filtersButtonDisabled).toBeDisabled()
-    expect(screen.getByTestId('filters-label')).toHaveTextContent('Pending filter activated: 1 patients')
+    expect(screen.getByTestId('filters-label')).toHaveTextContent('1 pending patient(s)')
     expect(screen.queryByTestId('reset-filters-link')).not.toBeInTheDocument()
 
     expect(within(dataGridRow).getAllByRole('row')).toHaveLength(2)